let user store objects in the session
By default, Session::Storage::Secure does not allow storing objects
due to concerns of action-at-a-distance. This is because, as
Sereal::Encoder's documentation puts it, blessed references can mean
executing a destructor on a remote system or generally executing code
based on data. In a web application, however, it is very common to
store JSON data in sessions (like when handling oauth from third-party
providers). The Perl implementation of JSON creates thin "JSON::bool"
objects to make the 1/true 0/false relationship between JSON data and
Perl data structures, but you can't store it in Dancer2::Session::Cookie
because it doesn't allow objects. Besides, since session handling is
performed by the developer, it should be the developer's responsibility
to either store objects or not on the session structure.

This patch fixes the issues above.