Add noop _change_id method ready for future change_session_id in D2
A future Dancer2 release is expected to include a `change_session_id`
keyword which expects session engines to change the session ID via a
supplied `_change_id` method. For this session engine change_id is a
noop since ID changes on each request anyway so add it along with tests
anyway to prevent possible future regression.